Fix the last bit of needed code duplication
